Split main.cpp into smaller files
