+++ Makefile
@@ -17,8 +17,8 @@ NETWORKING=$(PROJECT_DIR)/networking
 .PHONY: local_dns
 .PHONY: install
 
-main: $(NETWORKING)/utils.global.o $(NETWORKING)/ip_getter.global.o $(NETWORKING)/crafter_requester.global.o $(CONFIG)/DnsMap.global.o $(CONFIG)/DnsMapCache.global.o $(CONFIG)/DnsMapUser.global.o $(CONFIG)/DnsMapUserSettings.global.o $(LIB)/time_utils.o
-	g++ -std=c++20 -L/usr/local/lib -Wl,-rpath=/usr/local/lib -DGLOBAL=1 $(DNS_SERVER)/main.cpp -o $(DNS_SERVER)/main.x $(NETWORKING)/utils.global.o $(NETWORKING)/ip_getter.global.o $(NETWORKING)/crafter_requester.global.o $(CONFIG)/DnsMap.global.o $(CONFIG)/DnsMapCache.global.o $(CONFIG)/DnsMapUser.global.o $(CONFIG)/DnsMapUserSettings.global.o $(LIB)/time_utils.o -pthread -lcrafter -lnsl -lrt -lpcap -lm -lresolv
+main: $(NETWORKING)/utils.global.o $(NETWORKING)/ip_getter.global.o $(NETWORKING)/crafter_requester.global.o $(CONFIG)/DnsMap.global.o $(CONFIG)/DnsMapCache.global.o $(CONFIG)/DnsMapUser.global.o $(CONFIG)/DnsMapUserSettings.global.o $(LIB)/time_utils.o $(DNS_SERVER)/dns_server_utils.global.o $(DNS_SERVER)/tcp.global.o $(DNS_SERVER)/udp.global.o 
+	g++ -std=c++20 -L/usr/local/lib -Wl,-rpath=/usr/local/lib -DGLOBAL=1 $(DNS_SERVER)/main.cpp -o $(DNS_SERVER)/main.x $(NETWORKING)/utils.global.o $(NETWORKING)/ip_getter.global.o $(NETWORKING)/crafter_requester.global.o $(CONFIG)/DnsMap.global.o $(CONFIG)/DnsMapCache.global.o $(CONFIG)/DnsMapUser.global.o $(CONFIG)/DnsMapUserSettings.global.o $(LIB)/time_utils.o $(DNS_SERVER)/dns_server_utils.global.o $(DNS_SERVER)/tcp.global.o $(DNS_SERVER)/udp.global.o -pthread -lcrafter -lnsl -lrt -lpcap -lm -lresolv
 
 %.global.o : %.cpp
 	g++ -DGLOBAL=1 -std=c++20 -I$(LIB_CRAFTER_SRC) -c $^ -o $@
