+++ networking/ip_getter.h
@@ -21,12 +21,12 @@ private:
 
     int timeout; //in miliseconds
 
-    DnsMapCache dnsMapCache;
+    DnsMapCache *dnsMapCache;
 
     std::string wait_for_promise(std::promise<std::string>& promise);
 public:
     std::string get_ip(std::string mac, int cacheTimeout); //This is the API the rest of the program has to call in order to invoke local-dns for a certain mac address
     
-    IPGetter(CrafterRequester *requester, DnsMapUser& dnsMapUser, std::string local_network_ip_mask1="192.168.0.*", int timeout1=1000);
+    IPGetter(CrafterRequester *requester, DnsMapCache *dnsMapCache, std::string local_network_ip_mask, int timeout);
 };
